requirements.txt
matplotlib==3.3.3
numpy==1.19.2
opencv_python==4.4.0.46
Pillow==9.2.0
scikit_learn==1.1.2
timm==0.6.7
torch==1.8.0a0+1606899
torchsummary==1.5.1
torchvision==0.9.0a0
tqdm==4.53.0
